+++ app/static/css/custom.css
