+++ app/templates/home.html
@@ -1,0 +1,44 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <!-- CSS -->
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css" integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
+
+    <!--Custom styles for this template -->
+    <link rel="stylesheet" href="{{ url_for('static', filename='css/custom.css') }}" >
+    <!-- JQuery and JS -->
+    <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>
+    <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.1/dist/umd/popper.min.js" integrity="sha384-9/reFTGAW83EW2RDu2S0VKaIzap3H66lZH81PoYlFhbGU+6BZp6G7niu735Sk7lN" crossorigin="anonymous"></script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/js/bootstrap.min.js" integrity="sha384-w1Q4orYjBQndcko6MimVbzY0tgp4pWB4lZ7lr30WKz0vr/aWKhXdBNmNb5D92v7s" crossorigin="anonymous"></script>
+
+    <title>Company Tracker</title>
+</head>
+<body>
+
+    <div class="jumbotron">
+        <h1 class="display-4">Welcome to the Company Tracker</h1>
+        <p class="lead">This is the Daily News Blog for the company Virgin Galactic. The news are retrieved from 
+            the NewsApi developed by Google. Please enter a company name and an email address. 
+            After the appplication is executed, an email will be sent and a podcast will be generated to listen to
+            the summary of the news.</p>
+       </div>
+
+
+       <form method="post">
+        <div class="form-group">
+          <label for="exampleInputEmail1">Email address</label>
+          <input type="email" class="form-control" name=email id="exampleInputEmail1" placeholder="name@example.com" aria-describedby="emailHelp">
+          <small id="email" class="form-text text-muted">We'll never share your email with anyone else.</small>
+        </div>
+        <div class="form-group">
+          <label for="exampleCompanyName">Company Name</label>
+          <input type="text" name=company_name class="form-control" id="company_name">
+        </div>
+        <button type="submit" class="btn btn-primary">Submit</button>
+      </form>
+
+    
+
+</body>
+</html>
